core-agent/requirements.txt
weaviate-client
pandas
requests
tqdm
scikit-learn
fastapi[all]
sqlalchemy
uvicorn
langchain
langgraph
langchain-google-genai
langchain-openai
pydantic
pymysql
html2text
rq
redis
argon2-cffi
PyJWT

nltk
sentence-transformers
torch
python-dotenv